feature: support PSE, Cash, and CC payments
In this commit, PSE, Cash, and Credit cards payments, has been completely re factored.
This code do the same as previous one, but using ES6 latest features (like classes over prototypes),
just CC payments had passed tests (manual test, generating payments)
