+++ .gitignore
@@ -1,2 +1,3 @@
+.vscode/
 node_modules/
-tests/testing.js
+tests/
