+++ lib/resources/cash.js
@@ -1,53 +1,70 @@
-/**
- * Module dependencies
- */
-var util = require('util');
-var EpaycoError = require('./errors');
-var Resource = require('./');
+'use strict'
 
-/**
- * Expose constructor
- */
-module.exports = cash;
+const Epayco = require('../Epayco')
 
-/**
- * Customers constructor
- */
-function cash(epayco) {
-    Resource.call(this, epayco);
+class Cash extends Epayco {
+  constructor ({ apiKey, privateKey, lang, test }) {
+    super({ apiKey, privateKey, lang, test })
+  }
+
+  async create ({ description, docNumber, docType, email, lastName, name, value, method }) {
+    try {
+      const paymentDefaults = {
+        currency: 'COP',
+        method_confirmation: 'GET',
+        tax: '0',
+        tax_base: '0',
+        url_confirmation: 'https:/secure.payco.co/restpagos/testRest/endpagopse.php',
+        url_response: 'https:/secure.payco.co/restpagos/testRest/endpagopse.php'
+      }
+
+      const paymentData = {
+        ...paymentDefaults,
+        description,
+        doc_number: docNumber.toString(),
+        doc_type: docType.toString(),
+        email: email.toString(),
+        end_date: '2019-10-10', // MOMENTJS
+        last_name: lastName.toString(),
+        name: name.toString(),
+        value: value.toString()
+      }
+
+      const apiUrl = this.getCashAPIUrlOf(method)
+      const createCashPromise = this.request('POST', apiUrl, paymentData, true)
+      const cash = await createCashPromise
+      return cash
+    } catch (err) {
+      throw err
+    }
+  }
+
+  // TODO: (juanarbol)
+  // This is a duplication of charge.get() method
+  // Apply DRY in this
+  async get (uid) {
+    try {
+      const apiUrl = `/restpagos/transaction/response.json?ref_payco=${uid}&&public_key=${this.apiKey}`
+      const getChargePromise = this.request('GET', apiUrl, {}, true)
+      const chargeInfo = await getChargePromise
+      return chargeInfo
+    } catch (err) {
+      throw err
+    }
+  }
+
+  getCashAPIUrlOf (method) {
+    switch (method) {
+      case 'Baloto':
+        return '/restpagos/pagos/balotos.json'
+      case 'Efecty':
+        return '/restpagos/pagos/efecties.json'
+      case 'Gana':
+        return '/restpagos/pagos/ganas.json'
+      default:
+        return '/restpagos/pagos/efecties.json'
+    }
+  }
 }
 
-util.inherits(cash, Resource);
-
-/**
- * Create Subscriptions
- * @param {Object} options
- * @api public
- */
-cash.prototype.create = function(type, options) {
-    var url;
-    switch (type) {
-        case "efecty":
-            url = "/restpagos/pagos/efecties.json";
-            break;
-        case "baloto":
-            url = "/restpagos/pagos/balotos.json";
-            break;
-        case "gana":
-            url = "/restpagos/pagos/ganas.json";
-            break;
-        default:
-            throw new EpaycoError(this._epayco.lang, 109);
-    }
-    return this.request('post', url, options, sw = true);
-};
-
-/**
- * Retrieve Subscriptions
- *
- * @param {String} uid
- * @api public
- */
-cash.prototype.get = function(uid) {
-    return this.request('get', "/restpagos/transaction/response.json?ref_payco=" + uid + "&&public_key=" + this._epayco.apiKey, {}, sw = true);
-};
+module.exports = Cash
