+++ lib/resources/charge.js
@@ -1,4 +1,6 @@
-const Epayco = require('../index')
+'use strict'
+
+const Epayco = require('../Epayco')
 
 class Charge extends Epayco {
   constructor ({ apiKey, privateKey, lang, test }) {
