+++ lib/resources/token.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const Epayco = require('../index')
+const Epayco = require('../Epayco')
 const EpaycoError = require('../EpaycoError')
 
 class Token extends Epayco {
