--- tests/test.js
@@ -1,12 +1,0 @@
-describe('Epayco', function() {
-    describe('Constructor', function() {
-        it('Refuse to initialize without an apikey', function(done) {
-            try {
-                var epayco = Epayco();
-            } catch (e) {
-                assert(e);
-                done();
-            }
-        });
-    });
-});
